+++ CHANGELOG.md
@@ -15,7 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - added average secrets per completion
 - Dungeon Party Finder: customizable Party Notes filters (`/moo dungeon rules` or `/moo dr`)
   - Add and edit additional rules to mark Dungeon parties based on their Party notes
-- Added data for Enderman slayer, Voidling minions, and Hard Stone minions
+- Added data for Enderman slayer, Voidling minions, Hard Stone minions, and Golden Dragon (lvl 101+ pets)
 - New keybindings to...
   1) run `/moo waila` command (disabled by default; MC Options > Controls > `Cowlection`)
   2) copy a single item to clipboard as JSON with <kbd>CTRL</kbd> + <kbd>C</kbd>